backend/positions/views.py
from django.shortcuts import render
from rest_framework import generics
from .models import Position
from .serializers import PositionSerializer


class PositionList(generics.ListCreateAPIView):
    queryset = Position.objects.all()
    serializer_class = PositionSerializer


class PositionDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Position.objects.all()
    serializer_class = PositionSerializer
